requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
configparser==5.0.2
cvxpy==1.1.17
cycler==0.11.0
docker-pycreds==0.4.0
ecos==2.0.7.post1
einops==0.3.2
filelock==3.3.2
frozenlist==1.2.0
fsspec==2021.10.1
future==0.18.2
gdown==4.2.0
gitdb==4.0.9
gitpython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
kiwisolver==1.3.2
lightning-bolts==0.3.0
llvmlite==0.39.1
markdown==3.3.4
matplotlib==3.4.3
multidict==5.2.0
numba==0.56.2
oauthlib==3.1.1
opencv-python==4.5.4.58
osqp==0.6.2.post0
pandas==1.3.5
pathtools==0.1.2
pillow-simd==7.0.0.post3
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynndescent==0.5.7
pyparsing==3.0.4
pysocks==1.7.1
pytorch-lightning==1.1.3
pytz==2021.3
pyyaml==6.0
qdldl==0.1.5.post0
qpth
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
scs==2.1.4
seaborn==0.11.2
sentry-sdk==1.4.3
shortuuid==1.0.1
sklearn==0.0
smmap==5.0.0
soupsieve==2.2.1
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardx==2.4.1
termcolor==1.1.0
thop==0.0.31-2005241907
threadpoolctl==3.0.0
timm==0.4.12
tqdm==4.62.3
urllib3==1.26.7
wandb==0.12.7
werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.8.1
